web_socket_test/scripts/requirements.txt
#locustio==0.7.2
gevent-websocket==0.9.4
pyzmq==14.6.0
websocket-client==0.31.0
